small bug fixed

+++ src/EventSubscriber/CheckVerifiedUserSubscriber.php
@@ -29,7 +29,7 @@ class CheckVerifiedUserSubscriber implements EventSubscriberInterface
             );
         }
 
-        if (strlen($user->getGoogleAuthenticatorSecret()) > 0 && is_null($user->getTwoFaVerified())) {
+        if (!is_null($user->getGoogleAuthenticatorSecret()) && strlen($user->getGoogleAuthenticatorSecret()) > 0 && is_null($user->getTwoFaVerified())) {
             $this->accountService->removeTwofaRecoveryTokens($user);
             $this->accountService->unsetTwoStepVerification($user);
         }
